refactor: Consolidate comparison methods in ArrivalParetoSetComparatorFactory
